Extract actions from ApiRequestHandler
